refactor: move modules to `minimal_animatediff` directory
